To retrieve the channel we need id as a param, not for user
After attempting to consume YouTubeDataProvider.LoadChannelAsync( string ChannelID, int Size) I noticed that I wasn't getting a result. I tracked the issue to line 232. We are passing in fromUser which is not correct in this context. We want to pass in the param id. This in fact does return the correct data.